Implemented crude CLI command for executing the Advent Challenges
This so I don't have to add a new script for each Advent of Code challenge day.

I didn't use a CLI package here since I want the application to exit once the challenge day has been executed. This is so I don't have to exit a CLI so the code can be recompiled following refactors, then launch the CLI again to be able to trigger challenges.